wmendezblog/requirements.txt
decorator==4.0.11
Django==1.10.5
django-extensions==1.7.5
django-filter==1.0.1
django-simple-history==1.8.2
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
drfdocs==0.0.11
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
PyJWT==1.4.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
Werkzeug==0.11.15
